Move printClassNames to TastyPrinter
